+++ admin-ui/cspell.json
@@ -1,0 +1,100 @@
+{
+  "version": "0.2",
+  "language": "en",
+  "ignorePaths": [
+    "node_modules/**",
+    "dist/**",
+    "coverage/**",
+    "package-lock.json",
+    "jans_config_api_orval/**",
+    "configApiSpecs.yaml"
+  ],
+  "words": [
+    "acrs",
+    "adminui",
+    "agama",
+    "axios",
+    "axiosInstance",
+    "babel",
+    "browserslist",
+    "cedarling",
+    "cedarlings",
+    "cedarpolicy",
+    "configapi",
+    "createrole",
+    "datepicker",
+    "dayjs",
+    "devtools",
+    "dnd",
+    "dotenv",
+    "esbuild",
+    "fido",
+    "formik",
+    "gluu",
+    "gluuserver",
+    "hasher",
+    "hmr",
+    "idtoken",
+    "iframe",
+    "inum",
+    "jans",
+    "jansAdminUIRole",
+    "janssen",
+    "janssenproject",
+    "javascripts",
+    "jsons",
+    "jszip",
+    "jwks",
+    "jwt",
+    "keycloak",
+    "lockreducer",
+    "logoutaudit",
+    "mau",
+    "memberof",
+    "moduleresolution",
+    "noopener",
+    "noreferrer",
+    "ngrok",
+    "oauth",
+    "oidc",
+    "openapi",
+    "openid",
+    "openidconnect",
+    "orval",
+    "passwordless",
+    "persistor",
+    "persistors",
+    "postcss",
+    "rehydrate",
+    "rehydrated",
+    "resourceids",
+    "saml",
+    "scim",
+    "scopeid",
+    "scopereducer",
+    "scriptid",
+    "sessionauth",
+    "sessionrecovery",
+    "smtp",
+    "ssa",
+    "ssas",
+    "subgraph",
+    "tanstack",
+    "tooltip",
+    "toastify",
+    "tsdk",
+    "tsx",
+    "typeahead",
+    "tss",
+    "ujwt",
+    "userinfo",
+    "userinum",
+    "vitejs",
+    "wasm",
+    "webpack",
+    "webhook",
+    "webhooks",
+    "yaml"
+  ],
+  "flagWords": []
+}
