+++ admin-ui/orval.config.ts
@@ -12,10 +12,10 @@ export default defineConfig({
       client: 'react-query',
       httpClient: 'axios',
       mock: false,
-      prettier: true,
+      formatter: 'prettier',
       override: {
         mutator: {
-          path: './orval/orvalMutator.ts',
+          path: './orval/axiosInstance.ts',
           name: 'customInstance',
         },
         query: {
