+++ admin-ui/orval/index.ts
@@ -1,1 +1,3 @@
-export { customInstance, AXIOS_INSTANCE } from './orvalMutator'
+export { AXIOS_INSTANCE, setApiToken, customInstance } from './axiosInstance'
+export { installInterceptors } from './interceptors'
+export type { CancellablePromise, RetriableAxiosRequestConfig } from './types'
