+++ admin-ui/tsconfig.json
@@ -22,14 +22,12 @@
       "Routes/*": ["./app/routes/*"],
       "Plugins/*": ["./plugins/*"],
       "Context/*": ["./app/context/*"],
-      "context/*": ["./app/context/*"],
       "Utils/*": ["./app/utils/*"],
       "Styles/*": ["./app/styles/*"],
       "Orval": ["./orval/index"],
       "Orval/*": ["./orval/*"],
       "Redux/*": ["./app/redux/*"],
       "Images/*": ["./app/images/*"],
-      "SVG/*": ["./app/components/SVG/*"],
       "JansConfigApi": ["./jans_config_api_orval/src/JansConfigApi"]
     },
     "outDir": "./dist",
